requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-jose
pytest
httpx
requests
bcrypt
supabase==2.9.1
realtime==2.0.6
gotrue==2.9.3
storage3==0.8.2
postgrest==0.17.2
httpx==0.27.2
supafunc==0.6.2
aiohttp==3.10.10
python-dotenv
beautifulsoup4
lxml
google-generativeai